Move args into std::sys
